Move more composer code to separate files.
